just added some comments for better understanding

+++ physics-engine/Transform.h
@@ -9,7 +9,8 @@
 
 struct Transform
 {
-	glm::vec3 position = glm::vec3(0.0f);
+	//creating a structure to use for the positon,rotation,and scale of the object
+	glm::dvec3 position = glm::dvec3(0.0f);
 	glm::quat rotation = glm::quat(1.0f, 0.0f, 0.0f, 0.0f);
 	glm::vec3 scale = glm::vec3(1.0f);
